requirements.txt
numpy
PyOpenGL
noise
pygame